+++ src/main/kotlin/com/github/rinchinov/ijdbtplugin/CopyPasteActionsInterface.kt
@@ -1,13 +1,14 @@
 package com.github.rinchinov.ijdbtplugin
 
-import com.intellij.openapi.actionSystem.ActionManager
+import com.github.rinchinov.ijdbtplugin.services.Notifications
+import com.intellij.notification.NotificationType
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.CommonDataKeys
 import com.intellij.openapi.application.ApplicationManager
 import com.intellij.openapi.command.WriteCommandAction
+import com.intellij.openapi.components.service
 import com.intellij.openapi.editor.Editor
 import com.intellij.openapi.ide.CopyPasteManager
-import com.intellij.openapi.project.Project
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
 import java.awt.datatransfer.DataFlavor
@@ -18,17 +19,27 @@ interface CopyPasteActionsInterface {
 
     fun replaceRefsAndSourcesFromJinja2(query: String, target: String): String
     fun replaceRefsAndSourcesToJinja2(query: String, target: String): String
-    fun copyWithReplacingRefsAndSources(e: AnActionEvent, target: String){
-        coroutineScope.launch {
+    fun getWithReplacingRefsAndSources(e: AnActionEvent, target: String): String {
+            var replacedContentResult = ""
             ApplicationManager.getApplication().runReadAction {
                 val editor: Editor = e.getRequiredData(CommonDataKeys.EDITOR)
                 val document = editor.document
                 val selectionModel = editor.selectionModel
                 val selectedText = selectionModel.selectedText ?: document.text
-                val replacedContent = replaceRefsAndSourcesFromJinja2(selectedText, target)
-                val copyPasteManager = CopyPasteManager.getInstance()
-                copyPasteManager.setContents(StringSelection(replacedContent))
+                replacedContentResult = replaceRefsAndSourcesFromJinja2(selectedText, target)
             }
+            return replacedContentResult
+        }
+    fun copyWithReplacingRefsAndSources(e: AnActionEvent, target: String){
+        coroutineScope.launch {
+            val replacedContentResult = getWithReplacingRefsAndSources(e, target)
+            val copyPasteManager = CopyPasteManager.getInstance()
+            copyPasteManager.setContents(StringSelection(replacedContentResult))
+            e.project?.service<Notifications>()?.sendNotification(
+                "Copied with replaced refs/sources",
+                "",
+                NotificationType.INFORMATION
+            )
         }
     }
     fun pasteWithReplacedRefsAndSources(e: AnActionEvent, target: String){