+++ src/main/kotlin/com/github/rinchinov/ijdbtplugin/extensions/PluginProjectConfigurable.kt
@@ -12,7 +12,7 @@ import javax.swing.*
 import com.intellij.ui.components.JBScrollPane
 
 
-class PluginProjectConfigurable(private val project: Project) : Configurable {
+class PluginProjectConfigurable(project: Project) : Configurable {
     private val settings = project.service<ProjectSettings>()
     private val projectConfigurations = project.service<ProjectConfigurations>()
     private val mainPanel = JPanel(GridBagLayout())
@@ -20,9 +20,10 @@ class PluginProjectConfigurable(private val project: Project) : Configurable {
     private val dbtProfileDir: JTextField = JTextField()
     private val dbtRunnerImport: JTextField = JTextField()
     private val dbtInterpreterPath: JTextField = JTextField()
-    private val dbtTargetList: JTextField = JTextField()
-    private val dbtDefaultTarget: JTextField = JTextField()
-    private val dbtAdapter: JTextField = JTextField()
+    private val dbtQueryRunPaginationTemplate: JTextField = JTextField()
+    private val dbtQueryRunCountTemplate: JTextField = JTextField()
+    private val dbtQueryRunDryTemplate: JTextField = JTextField()
+    private val dbtQueryRunPlanTemplate: JTextField = JTextField()
 
     init {
         val gbc = GridBagConstraints().apply {
@@ -37,24 +38,27 @@ class PluginProjectConfigurable(private val project: Project) : Configurable {
         addLabeledField("DBT profile path", dbtProfileDir, mainPanel, gbc)
         addLabeledField("DBT runner import", dbtRunnerImport, mainPanel, gbc)
         addLabeledField("DBT interpreter path", dbtInterpreterPath, mainPanel, gbc)
-        addLabeledField("DBT target lists", dbtTargetList, mainPanel, gbc)
-        addLabeledField("DBT plugin default target", dbtDefaultTarget, mainPanel, gbc)
-        addLabeledField("DBT plugin default adapter", dbtAdapter, mainPanel, gbc)
 
+        addLabeledField("Query run SQL templates:", null, mainPanel, gbc)
+        addLabeledField("Query paginated template", dbtQueryRunPaginationTemplate, mainPanel, gbc)
+        addLabeledField("Query count template", dbtQueryRunCountTemplate, mainPanel, gbc)
+        addLabeledField("Query plan template", dbtQueryRunPlanTemplate, mainPanel, gbc)
+        addLabeledField("Query dry run template", dbtQueryRunDryTemplate, mainPanel, gbc)
         gbc.weighty = 1.0
         mainPanel.add(Box.createVerticalGlue(), gbc)
     }
 
-    private fun addLabeledField(labelText: String, field: JTextField, panel: JPanel, gbc: GridBagConstraints) {// Clone the GridBagConstraints to avoid side effects
+    private fun addLabeledField(labelText: String, field: JTextField?, panel: JPanel, gbc: GridBagConstraints) {// Clone the GridBagConstraints to avoid side effects
         val labelGbc = gbc.clone() as GridBagConstraints
         labelGbc.gridx = 0
         labelGbc.weightx = 0.0
         panel.add(JLabel(labelText), labelGbc)
-
-        val fieldGbc = gbc.clone() as GridBagConstraints
-        fieldGbc.gridx = 1
-        fieldGbc.weightx = 1.0
-        panel.add(field, fieldGbc)
+        if (field!=null){
+            val fieldGbc = gbc.clone() as GridBagConstraints
+            fieldGbc.gridx = 1
+            fieldGbc.weightx = 1.0
+            panel.add(field, fieldGbc)
+        }
         gbc.gridy += 2
     }
 
@@ -67,9 +71,10 @@ class PluginProjectConfigurable(private val project: Project) : Configurable {
                 || settings.getDbtProfileDir() != dbtProfileDir.text
                 || settings.getDbtRunnerImport() != dbtRunnerImport.text
                 || settings.getDbtInterpreterPath() != dbtInterpreterPath.text
-                || settings.getDbtDefaultTarget() != dbtDefaultTarget.text
-                || settings.getDbtAdapter() != dbtAdapter.text
-                || settings.getDbtTargetList() != dbtTargetList.text.split(",")
+                || settings.getDbtQueryRunPlanTemplate() != dbtQueryRunPaginationTemplate.text
+                || settings.getDbtQueryRunCountTemplate() != dbtQueryRunCountTemplate.text
+                || settings.getDbtQueryRunPlanTemplate() != dbtQueryRunPlanTemplate.text
+                || settings.getDbtQueryRunDryTemplate() != dbtQueryRunDryTemplate.text
     }
 
     override fun apply() {
@@ -77,9 +82,10 @@ class PluginProjectConfigurable(private val project: Project) : Configurable {
         settings.setDbtProfileDir(dbtProfileDir.text)
         settings.setDbtRunnerImport(dbtRunnerImport.text)
         settings.setDbtInterpreterPath(dbtInterpreterPath.text)
-        settings.setDbtDefaultTarget(dbtDefaultTarget.text)
-        settings.setDbtAdapter(dbtAdapter.text)
-        settings.setDbtTargetList(dbtTargetList.text)
+        settings.setDbtQueryRunPaginationTemplate(dbtQueryRunPaginationTemplate.text)
+        settings.setDbtQueryRunCountTemplate(dbtQueryRunCountTemplate.text)
+        settings.setDbtQueryRunDryTemplate(dbtQueryRunDryTemplate.text)
+        settings.setDbtQueryRunPlanTemplate(dbtQueryRunPlanTemplate.text)
         projectConfigurations.reloadDbtProjectSettings()
     }
 
@@ -88,13 +94,12 @@ class PluginProjectConfigurable(private val project: Project) : Configurable {
         dbtProfileDir.text = settings.getDbtProfileDir()
         dbtRunnerImport.text = settings.getDbtRunnerImport()
         dbtInterpreterPath.text = settings.getDbtInterpreterPath()
-        dbtDefaultTarget.text = settings.getDbtDefaultTarget()
-        dbtAdapter.text = settings.getDbtAdapter()
-        dbtTargetList.text = settings.getDbtTargetList().joinToString(separator = ",")
+        dbtQueryRunPaginationTemplate.text = settings.getDbtQueryRunPaginationTemplate()
+        dbtQueryRunCountTemplate.text = settings.getDbtQueryRunCountTemplate()
+        dbtQueryRunDryTemplate.text = settings.getDbtQueryRunDryTemplate()
+        dbtQueryRunPlanTemplate.text = settings.getDbtQueryRunPlanTemplate()
     }
 
     override fun getDisplayName(): String = MyBundle.message("settingWindowName")
-    override fun disposeUIResources() {
-
-    }
+    override fun disposeUIResources() {}
 }