+++ src/main/kotlin/com/github/rinchinov/ijdbtplugin/queryExecution/QueryChangeListener.kt
@@ -1,0 +1,8 @@
+package com.github.rinchinov.ijdbtplugin.queryExecution
+
+interface QueryChangeListener {
+    fun onQueryAdd(queryExecution: QueryExecutionBackend.QueryExecution, keep: Int)
+    fun displayExecutedQuery(queryExecution: QueryExecutionBackend.QueryExecution, pageNumber: Int)
+    fun displayQueryLoading()
+    fun clearQueryPanel()
+}