+++ src/main/kotlin/com/github/rinchinov/ijdbtplugin/services/ProjectConfigurations.kt
@@ -1,5 +1,6 @@
 package com.github.rinchinov.ijdbtplugin.services
 
+import com.github.rinchinov.ijdbtplugin.extensions.MainToolWindowService
 import com.github.rinchinov.ijdbtplugin.utils.renderJinjaEnvVar
 import com.intellij.notification.NotificationType
 import com.intellij.openapi.components.Service
@@ -19,13 +20,17 @@ class ProjectConfigurations(private val project: Project) {
     val settings = project.service<ProjectSettings>()
     private val dbtNotifications = project.service<Notifications>()
     private val eventLoggerManager = project.service<EventLoggerManager>()
-    private val dbtProjectConfig = DbtProjectConfig(
+    val dbtProjectConfig = DbtProjectConfig(
         "",
         1,
         "",
         "",
         "target",
-        "dbt_packages"
+        "dbt_packages",
+        null,
+        null,
+        emptyList(),
+        null
     )
     data class DbtProjectConfig(
         var name: String,
@@ -33,10 +38,27 @@ class ProjectConfigurations(private val project: Project) {
         var version: String,
         var profile: String,
         var targetPath: String,
-        var packagesInstallPath: String
+        var packagesInstallPath: String,
+        var projectProfiles: Map<String, Map<String, Any>>?,
+        var defaultTarget: String?,
+        var targets: List<String>,
+        var adapterName: String?
     )
     init {
         reloadDbtProjectSettings()
+    }
+    fun getProfileDetails(target: String): Map<String, Any>? {
+        return dbtProjectConfig.projectProfiles?.get(target)
+    }
+    private fun loadProfileDetails(){
+        val profileFile = Paths.get(getDbtProfileDirAbsolute().toString(), "profiles.yml")
+        val inputStream: InputStream = Files.newInputStream(profileFile)
+        val profilesRaw = Yaml().load(inputStream) as Map<String, Map<String, Map<String, Map<String, Any>>>>?
+        val raw = profilesRaw?.get(dbtProjectConfig.name)
+        dbtProjectConfig.projectProfiles = raw?.get("outputs") as Map<String, Map<String, Any>>
+        dbtProjectConfig.defaultTarget = raw["target"] as String
+        dbtProjectConfig.targets = dbtProjectConfig.projectProfiles?.keys?.toList() ?: emptyList()
+        dbtProjectConfig.adapterName = dbtProjectConfig.projectProfiles!![dbtProjectConfig.defaultTarget!!]?.get("type") as String
     }
     fun reloadDbtProjectSettings(){
         val filePath = dbtProjectPath().absolutePath.toString()
@@ -54,14 +76,17 @@ class ProjectConfigurations(private val project: Project) {
                     dbtProjectConfig.packagesInstallPath
                 ) as String
                 dbtProjectConfig.packagesInstallPath = renderJinjaEnvVar(packagesInstallPath)
+                loadProfileDetails()
             }
             else {
                 dbtNotifications.sendNotification("Load project failed", filePath, NotificationType.ERROR)
             }
         } catch (e: FileNotFoundException) {
-            dbtNotifications.sendNotification("File not found", "$filePath\n TBD Instruction and link to the doc", NotificationType.ERROR)
+            eventLoggerManager.logLines(e.stackTraceToString().lines(), "core")
+            dbtNotifications.sendNotification("File not found", "Failed to open `$filePath`", NotificationType.ERROR, MainToolWindowService.Tab.LOGS)
         } catch (e: Exception) {
-            dbtNotifications.sendNotification("Error loading YAML file", ": ${e.message}\nTBD Instruction and link to the doc", NotificationType.ERROR)
+            eventLoggerManager.logLines(e.stackTraceToString().lines(), "core")
+            dbtNotifications.sendNotification("Error loading YAML file", "Failed to open `$filePath`", NotificationType.ERROR, MainToolWindowService.Tab.LOGS)
         }
         eventLoggerManager.notifyProjectConfigurationsChangeListeners(this)
     }
@@ -130,4 +155,5 @@ class ProjectConfigurations(private val project: Project) {
         dbtNotifications.sendNotification("Can't find python interpreter", "", NotificationType.ERROR)
         return ""
     }
+
 }